docs/examples/py/requirements.txt
reduct-py @ https://github.com/reductstore/reduct-py/archive/refs/heads/main.zip
eclipse-zenoh~=1.7